backend/pyproject.toml
[tool.poetry]
name = "ezev"
version = "0.1.0"
description = ""
authors = ["denny wright <dnnywrght@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.0"
asyncpg = "0.24.0"
loguru = "^0.6.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
httpx = "^0.22.0"
tortoise-orm = {extras = ["accel"], version = "^0.18.1"}
aerich = "^0.6.2"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pytest = "^7.1.0"
pytest-cov = "^3.0.0"
pytest-sugar = "^0.9.4"
pytest-asyncio = "^0.18.2"
flake8 = "^4.0.1"
flake8-bugbear = "^22.1.11"
flake8-comprehensions = "^3.8.0"
flake8-eradicate = "^1.2.0"
flake8-functions-names = "^0.3.0"
flake8-pep3101 = "^1.3.0"
flake8-pytest-style = "^1.6.0"
flake8-return = "^1.1.3"
flake8-simplify = "^0.18.1"
flake8-variables-names = "^0.0.5"
flake8-cognitive-complexity = "^0.1.0"
flake8-functions = "^0.0.7"
python-dotenv = "^0.19.2"


[tool.aerich]
tortoise_orm = "src.utils.db_manager.MIGRATION_CONFIG"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )
)
'''